+++ public/pages/Main/Main.tsx
@@ -14,8 +14,6 @@ import {
   EuiTitle,
   EuiSpacer,
   EuiGlobalToastList,
-  EuiFlexGroup,
-  EuiFlexItem,
 } from "@elastic/eui";
 import { Toast } from "@opensearch-project/oui/src/eui_components/toast/global_toast_list";
 import { AppMountParameters, CoreStart } from "opensearch-dashboards/public";
@@ -104,6 +102,7 @@ import {
 import { GettingStartedContent } from "../Overview/components/GettingStarted/GettingStartedContent";
 import { BrowserServices } from "../../models/interfaces";
 import { CHANNEL_TYPES } from "../CreateDetector/components/ConfigureAlerts/utils/constants";
+import { DecoderFormPage } from "../Decoders/containers/DecoderFormPage";
 
 enum Navigation {
   SecurityAnalytics = "Security Analytics",
@@ -146,6 +145,8 @@ const HIDDEN_NAV_ROUTES: string[] = [
   ROUTES.THREAT_INTEL_ADD_CUSTOM_SOURCE,
   ROUTES.THREAT_INTEL_CREATE_SCAN_CONFIG,
   ROUTES.THREAT_INTEL_EDIT_SCAN_CONFIG,
+  ROUTES.DECODERS_CREATE,
+  ROUTES.DECODERS_EDIT,
 ];
 
 interface MainProps extends RouteComponentProps {
@@ -413,7 +414,9 @@ export default class Main extends Component<MainProps, MainState> {
               // this.setState({ selectedNavItemId: Navigation.Findings });
               // history.push(ROUTES.FINDINGS);
               // Wazuh: navigate to app so this is highlighted in the sidebar menu (old)
-              getApplication().navigateToApp(FINDINGS_NAV_ID, {path: generateAppPath(ROUTES.FINDINGS)});
+              getApplication().navigateToApp(FINDINGS_NAV_ID, {
+                path: generateAppPath(ROUTES.FINDINGS),
+              });
             },
             isSelected: selectedNavItemId === Navigation.Findings,
           },
@@ -1120,7 +1123,36 @@ export default class Main extends Component<MainProps, MainState> {
                                     <Route
                                       path={ROUTES.DECODERS}
                                       render={(props: RouteComponentProps) => (
-                                        <Decoders {...props} />
+                                        <Decoders
+                                          {...props}
+                                          notifications={core?.notifications}
+                                        />
+                                      )}
+                                    />
+                                    <Route
+                                      path={ROUTES.DECODERS_CREATE}
+                                      render={(props: RouteComponentProps) => (
+                                        <DecoderFormPage
+                                          {...props}
+                                          notifications={core?.notifications}
+                                          action="create"
+                                          history={props.history}
+                                        />
+                                      )}
+                                    />
+                                    <Route
+                                      path={`${ROUTES.DECODERS_EDIT}/:id`}
+                                      render={(
+                                        props: RouteComponentProps<{
+                                          id: string;
+                                        }>,
+                                      ) => (
+                                        <DecoderFormPage
+                                          {...props}
+                                          notifications={core?.notifications}
+                                          action="edit"
+                                          history={props.history}
+                                        />
                                       )}
                                     />
                                     <Route