+++ server/routes/DecodersRoutes.ts
@@ -1,13 +1,13 @@
 /*
  * Copyright Wazuh Inc.
  * SPDX-License-Identifier: AGPL-3.0-or-later
-*/
+ */
 
-import { IRouter } from 'opensearch-dashboards/server';
-import { schema } from '@osd/config-schema';
-import { NodeServices } from '../models/interfaces';
-import { API } from '../utils/constants';
-import { createQueryValidationSchema } from '../utils/helpers';
+import { IRouter } from "opensearch-dashboards/server";
+import { schema } from "@osd/config-schema";
+import { NodeServices } from "../models/interfaces";
+import { API } from "../utils/constants";
+import { createQueryValidationSchema } from "../utils/helpers";
 
 export function setupDecodersRoutes(services: NodeServices, router: IRouter) {
   const { decodersService } = services;
@@ -22,7 +22,7 @@ export function setupDecodersRoutes(services: NodeServices, router: IRouter) {
         }),
       },
     },
-    decodersService.searchDecoders
+    decodersService.searchDecoders,
   );
 
   router.get(
@@ -37,6 +37,65 @@ export function setupDecodersRoutes(services: NodeServices, router: IRouter) {
         }),
       },
     },
-    decodersService.getDecoder
+    decodersService.getDecoder,
+  );
+
+  router.post(
+    {
+      path: `${API.DECODERS_BASE}`,
+      validate: {
+        body: schema.object({
+          document: schema.any(),
+          integrationId: schema.string(),
+        }),
+        query: createQueryValidationSchema({
+          space: schema.maybe(schema.string()),
+        }),
+      },
+    },
+    decodersService.createDecoder,
+  );
+
+  router.put(
+    {
+      path: `${API.DECODERS_BASE}/{decoderId}`,
+      validate: {
+        params: schema.object({
+          decoderId: schema.string(),
+        }),
+        body: schema.object({
+          document: schema.any(),
+        }),
+        query: createQueryValidationSchema({
+          space: schema.maybe(schema.string()),
+        }),
+      },
+    },
+    decodersService.updateDecoder,
+  );
+
+  router.delete(
+    {
+      path: `${API.DECODERS_BASE}/{decoderId}`,
+      validate: {
+        params: schema.object({
+          decoderId: schema.string(),
+        }),
+        query: createQueryValidationSchema({
+          space: schema.maybe(schema.string()),
+        }),
+      },
+    },
+    decodersService.deleteDecoder,
+  );
+
+  router.get(
+    {
+      path: `${API.DECODERS_BASE}/integrations/draft`,
+      validate: {
+        query: createQueryValidationSchema(),
+      },
+    },
+    decodersService.getDraftIntegrations,
   );
 }